+++ activesupport/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Rails 4.2.5.rc1 (October 30, 2015) ##
+## Rails 4.2.5 (November 12, 2015) ##
 
 *   Fix `TimeWithZone#eql?` to properly handle `TimeWithZone` created from `DateTime`:
         twz = DateTime.now.in_time_zone